+++ src/array.rs
@@ -100,13 +100,6 @@ pub type ArrayShape = Vec<u64>;
 #[error("value must be non-zero")]
 pub struct NonZeroError;
 
-/// An alias for bytes which may or may not be available.
-///
-/// When a value is read from a store, it returns `MaybeBytes` which is [`None`] if the key is not available.
-/// A bytes to bytes codec only decodes `MaybeBytes` holding actual bytes, otherwise the bytes are propagated to the next decoder.
-/// An array to bytes partial decoder must take care of converting missing chunks to the fill value.
-pub type MaybeBytes = Option<Vec<u8>>;
-
 /// A Zarr array.
 ///
 /// ## Initilisation