+++ src/storage.rs
@@ -99,6 +99,24 @@ pub type AsyncReadableListableStorage = Arc<dyn AsyncReadableListableStorageTrai
 /// [`Arc`] wrapped asynchronous readable, writable and listable storage.
 pub type AsyncReadableWritableListableStorage = Arc<dyn AsyncReadableWritableListableStorageTraits>;
 
+/// An alias for bytes which may or may not be available.
+///
+/// When a value is read from a store, it returns `MaybeBytes` which is [`None`] if the key is not available.
+///
+/// A bytes to bytes codec only decodes `MaybeBytes` holding actual bytes, otherwise the bytes are propagated to the next decoder.
+/// An array to bytes partial decoder must take care of converting missing chunks to the fill value.
+pub type MaybeBytes = Option<Vec<u8>>;
+
+#[cfg(feature = "async")]
+/// An alias for [`bytes::Bytes`].
+pub type AsyncBytes = bytes::Bytes;
+
+#[cfg(feature = "async")]
+/// An alias for bytes which may or may not be available.
+///
+/// When a value is read from a store, it returns `MaybeAsyncBytes` which is [`None`] if the key is not available.
+pub type MaybeAsyncBytes = Option<AsyncBytes>;
+
 /// A [`StoreKey`] and [`ByteRange`].
 #[derive(Debug, Clone)]
 pub struct StoreKeyRange {